Added sampler test cases. Renamed sampler.py to single_table.py
